Added handling and function of Home and End keys.
There are multiple possible escape sequences the home and end key can
send, based on the OS and terminal emulator run by the user. This code
handles all those cases.

The home key moves the cursor to the left side of the screen.
The end key moves the cursor to the right side of the screen.